web_pages/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
mysql-connector-python==8.0.33
requests==2.31.0
python-dotenv==1.0.0
transformers==4.33.0
torch==2.0.1
werkzeug==3.0.0
gunicorn==22.0.0

